cordon_gateway/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0
pyyaml==6.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
psycopg2-binary==2.9.9
